refactor(connectors): structure package into paths, storage, validation, and auth modules
