+++ ax_cli/connectors/storage.py
@@ -1,0 +1,61 @@
+"""Atomic JSON persistence helpers for connector state."""
+
+from __future__ import annotations
+
+import copy
+import json
+import os
+import tempfile
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any
+
+from .constants import FILE_MODE_SECRET
+
+
+def utc_now_iso() -> str:
+    return datetime.now(timezone.utc).isoformat()
+
+
+def atomic_write_json(path: Path, payload: dict[str, Any], *, mode: int = FILE_MODE_SECRET) -> None:
+    path.parent.mkdir(parents=True, exist_ok=True)
+    tmp_path: Path | None = None
+    try:
+        with tempfile.NamedTemporaryFile(
+            "w",
+            encoding="utf-8",
+            dir=path.parent,
+            prefix=f".{path.name}.",
+            suffix=".tmp",
+            delete=False,
+        ) as tmp:
+            tmp_path = Path(tmp.name)
+            json.dump(payload, tmp, indent=2, sort_keys=True)
+            tmp.write("\n")
+            tmp.flush()
+            os.fsync(tmp.fileno())
+        assert tmp_path is not None
+        tmp_path.chmod(mode)
+        tmp_path.replace(path)
+    finally:
+        if tmp_path is not None and tmp_path.exists():
+            try:
+                tmp_path.unlink()
+            except OSError:
+                pass
+    try:
+        path.chmod(mode)
+    except OSError:
+        pass
+
+
+def read_json_object(path: Path, *, default: dict[str, Any]) -> dict[str, Any]:
+    if not path.exists():
+        return copy.deepcopy(default)
+    try:
+        raw = json.loads(path.read_text(encoding="utf-8"))
+    except (OSError, ValueError):
+        return copy.deepcopy(default)
+    if not isinstance(raw, dict):
+        return copy.deepcopy(default)
+    return raw
